Pixel perfect, asset changes, DrawCircleOutline()
The game renders at a fixed 16:9 aspect ratio with changing resolution.

The Main class has a DebugDraw bool which can be used for enabling square and newly added circle outlines.

Removed the black part of the exit icon and replaced the retry icon with a floppy disc from the delete save icon