Fading, events and countdown actor
The Countdown actor now uses a bigger SpriteFont version of the font containing only digit characters (plus hyphen and dot). It also stops counting down if the win or lose condition is met.

Changed the _actors list of the Scene to be a protected member instead of just private.

Scenes now have events (SceneEnded and FirstUpdate) which are used to correctly display the scene fade.

There is now a fade when the current scene resets or changes.

There now exist more ways of getting actors from scenes.