Added draw passes
Draw passes can be used to draw elements with different properties. For example, UI has its own layer over everything and does not follow the camera, the background may use a shader effect etc.

Actors are now expected to select one of the available passes in the Draw() function instead of using the formerly provided SpriteBatch.